split out the reporting - the code is getting too large
